+++ source/extensions/omni.isaac.orbit/test/assets/test_articulation.py
@@ -19,6 +19,7 @@ simulation_app = app_launcher.app
 """Rest everything follows."""
 
 import ctypes
+import torch
 import traceback
 import unittest
 
@@ -38,7 +39,7 @@ class TestArticulation(unittest.TestCase):
         # Create a new stage
         stage_utils.create_new_stage()
         # Simulation time-step
-        self.dt = 0.01
+        self.dt = 0.005
         # Load kit helper
         sim_cfg = sim_utils.SimulationCfg(dt=self.dt, device="cuda:0", shutdown_app_on_stop=False)
         self.sim = sim_utils.SimulationContext(sim_cfg)
@@ -48,7 +49,6 @@ class TestArticulation(unittest.TestCase):
         # stop simulation
         self.sim.stop()
         # clear the stage
-        self.sim.clear()
         self.sim.clear_instance()
 
     """
@@ -81,9 +81,6 @@ class TestArticulation(unittest.TestCase):
             # update robot
             robot.update(self.dt)
 
-        # Delete articulation
-        del robot
-
     def test_initialization_fixed_base(self):
         """Test articulation initialization for fixed base."""
         # Create articulation
@@ -110,8 +107,100 @@ class TestArticulation(unittest.TestCase):
             # update robot
             robot.update(self.dt)
 
-        # Delete articulation
-        del robot
+    def test_external_force_on_single_body(self):
+        """Test application of external force on the base of the robot."""
+
+        # Robots
+        robot_cfg = ANYMAL_C_CFG
+        robot_cfg.spawn.func("/World/Anymal_c/Robot_1", robot_cfg.spawn, translation=(0.0, -0.5, 0.65))
+        robot_cfg.spawn.func("/World/Anymal_c/Robot_2", robot_cfg.spawn, translation=(0.0, 0.5, 0.65))
+        # create handles for the robots
+        robot = Articulation(robot_cfg.replace(prim_path="/World/Anymal_c/Robot.*"))
+
+        # Play the simulator
+        self.sim.reset()
+
+        # Find bodies to apply the force
+        body_ids, _ = robot.find_bodies("base")
+        # Sample a large force
+        external_wrench_b = torch.zeros(robot.root_view.count, len(body_ids), 6, device=self.sim.device)
+        external_wrench_b[..., 1] = 1000.0
+
+        # Now we are ready!
+        for _ in range(5):
+            # reset root state
+            root_state = robot.data.default_root_state_w.clone()
+            root_state[0, :2] = torch.tensor([0.0, -0.5], device=self.sim.device)
+            root_state[1, :2] = torch.tensor([0.0, 0.5], device=self.sim.device)
+            robot.write_root_state_to_sim(root_state)
+            # reset dof state
+            joint_pos, joint_vel = robot.data.default_joint_pos, robot.data.default_joint_vel
+            robot.write_joint_state_to_sim(joint_pos, joint_vel)
+            robot.reset()
+            # apply force
+            robot.set_external_force_and_torque(
+                external_wrench_b[..., :3], external_wrench_b[..., 3:], body_ids=body_ids
+            )
+            # perform simulation
+            for _ in range(100):
+                # apply action to the robot
+                robot.set_joint_position_target(robot.data.default_joint_pos.clone())
+                robot.write_data_to_sim()
+                # perform step
+                self.sim.step(render=app_launcher.RENDER)
+                # update buffers
+                robot.update(self.dt)
+            # check condition that the robots have fallen down
+            self.assertTrue(robot.data.root_pos_w[0, 2].item() < 0.2)
+            self.assertTrue(robot.data.root_pos_w[1, 2].item() < 0.2)
+
+    def test_external_force_on_multiple_bodies(self):
+        """Test application of external force on the legs of the robot."""
+
+        # Robots
+        robot_cfg = ANYMAL_C_CFG
+        robot_cfg.spawn.func("/World/Anymal_c/Robot_1", robot_cfg.spawn, translation=(0.0, -0.5, 0.65))
+        robot_cfg.spawn.func("/World/Anymal_c/Robot_2", robot_cfg.spawn, translation=(0.0, 0.5, 0.65))
+        # create handles for the robots
+        robot = Articulation(robot_cfg.replace(prim_path="/World/Anymal_c/Robot.*"))
+
+        # Play the simulator
+        self.sim.reset()
+
+        # Find bodies to apply the force
+        body_ids, _ = robot.find_bodies(".*_SHANK")
+        # Sample a large force
+        external_wrench_b = torch.zeros(robot.root_view.count, len(body_ids), 6, device=self.sim.device)
+        external_wrench_b[..., 1] = 100.0
+
+        # Now we are ready!
+        for _ in range(5):
+            # reset root state
+            root_state = robot.data.default_root_state_w.clone()
+            root_state[0, :2] = torch.tensor([0.0, -0.5], device=self.sim.device)
+            root_state[1, :2] = torch.tensor([0.0, 0.5], device=self.sim.device)
+            robot.write_root_state_to_sim(root_state)
+            # reset dof state
+            joint_pos, joint_vel = robot.data.default_joint_pos, robot.data.default_joint_vel
+            robot.write_joint_state_to_sim(joint_pos, joint_vel)
+            robot.reset()
+            # apply force
+            robot.set_external_force_and_torque(
+                external_wrench_b[..., :3], external_wrench_b[..., 3:], body_ids=body_ids
+            )
+            # perform simulation
+            for _ in range(100):
+                # apply action to the robot
+                robot.set_joint_position_target(robot.data.default_joint_pos.clone())
+                robot.write_data_to_sim()
+                # perform step
+                self.sim.step(render=app_launcher.RENDER)
+                # update buffers
+                robot.update(self.dt)
+            # check condition
+            # since there is a moment applied on the robot, the robot should rotate
+            self.assertTrue(robot.data.root_ang_vel_w[0, 2].item() > 0.1)
+            self.assertTrue(robot.data.root_ang_vel_w[1, 2].item() > 0.1)
 
 
 if __name__ == "__main__":